requirements.txt
nltk
spacy
scikit-learn
pandas
numpy
streamlit
pdfminer.six
pdfminer.six
